% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASIS_EDIT_single.R
\name{NASIS_EDIT_single}
\alias{NASIS_EDIT_single}
\title{generate NASIS to EDIT (single) report}
\usage{
NASIS_EDIT_single(ecosite, output_directory = getwd())
}
\arguments{
\item{ecosite}{ecosite code of interest}

\item{output_directory}{output directory for report (optional, will default output to package folder)}
}
\value{
html report
}
\description{
generate NASIS to EDIT (single) report
}
