% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASIS_EDIT_single.R
\name{NASIS_EDIT_single}
\alias{NASIS_EDIT_single}
\title{generate NASIS to EDIT (single) report}
\usage{
NASIS_EDIT_single(ecosite_id, data_path, output_directory)
}
\arguments{
\item{ecosite_id}{ecosite code of interest}

\item{data_path}{path to ecosite_report.csv (generated from NASIS Report)}

\item{output_directory}{output directory for report}
}
\value{
interactive html report
}
\description{
generate NASIS to EDIT (single) report
}
